scraper/requirements.txt
scrapy 
scrapy-playwright
playwright
psycopg2-binary

